requirements.txt
asgiref==3.8.1
certifi==2024.7.4
distlib==0.3.8
Django==4.2
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.19.12
filelock==3.15.4
pillow==11.0.0
pipenv==2024.0.1
platformdirs==4.2.2
python-decouple==3.8
setuptools==72.1.0
sqlparse==0.5.3
virtualenv==20.26.3
whitenoise==6.8.2
